+++ gatsby-node.js
@@ -16,7 +16,7 @@ module.exports.onCreateNode = ({ node, actions }) => {
 
 module.exports.createPages = async ({ graphql, actions }) => {
    const { createPage } = actions
-   const blogTemplate = path.resolve('./src/templates/blog.js')
+   const blogTemplate = path.resolve('./src/templates/blog/index.jsx')
    const res = await graphql(`
       query {
          allMarkdownRemark {
