+++ package.json
@@ -10,11 +10,13 @@
     "serve": "gatsby serve"
   },
   "dependencies": {
+    "babel-plugin-styled-components": "^1.10.6",
     "dotenv": "^8.2.0",
     "gatsby": "^2.18.21",
     "gatsby-image": "^2.2.38",
     "gatsby-plugin-google-analytics": "^2.1.32",
     "gatsby-plugin-sharp": "^2.3.13",
+    "gatsby-plugin-styled-components": "^3.1.17",
     "gatsby-remark-images": "^3.1.42",
     "gatsby-remark-prismjs": "^3.3.30",
     "gatsby-remark-relative-images": "^0.2.3",
@@ -23,7 +25,8 @@
     "gatsby-transformer-sharp": "^2.3.12",
     "prismjs": "^1.18.0",
     "react": "^16.12.0",
-    "react-dom": "^16.12.0"
+    "react-dom": "^16.12.0",
+    "styled-components": "^4.4.1"
   },
   "devDependencies": {
     "env-cmd": "^10.0.1"
