+++ src/pages/blog/index.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Link, graphql, useStaticQuery } from 'gatsby'
 
-import Layout from '../components/Layout'
+import Layout from '../../sections/Layout'
 
 const BlogPage = () => {
    const data = useStaticQuery(graphql`
