+++ src/pages/code/index.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import Layout from '../components/Layout'
+import Layout from '../../sections/Layout'
 
 export default () => {
    return <Layout>Code</Layout>
